1.15.4 block fire new base class shared props
