+++ src/main/java/com/lothrazar/cyclicmagic/block/fire/BlockFireSafe.java
@@ -1,0 +1,56 @@
+/*******************************************************************************
+ * The MIT License (MIT)
+ * 
+ * Copyright (C) 2014-2018 Sam Bassett (aka Lothrazar)
+ * 
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ * 
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ * 
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ ******************************************************************************/
+package com.lothrazar.cyclicmagic.block.fire;
+
+import com.lothrazar.cyclicmagic.potion.PotionEffectRegistry;
+import net.minecraft.block.state.IBlockState;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.EntityLivingBase;
+import net.minecraft.entity.EnumCreatureType;
+import net.minecraft.entity.player.EntityPlayer;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.world.World;
+
+public class BlockFireSafe extends BlockFireBase {
+
+  private static final int FIRESECONDS = 10;
+
+  public BlockFireSafe() {
+    super();
+  }
+
+  @Override
+  public void onEntityCollidedWithBlock(World worldIn, BlockPos pos, IBlockState state, Entity entityIn) {
+    if (!worldIn.isRemote && entityIn instanceof EntityLivingBase
+        && !(entityIn instanceof EntityPlayer)) {
+      EntityLivingBase e = ((EntityLivingBase) entityIn);
+      if (!e.isPotionActive(PotionEffectRegistry.SNOW)
+          && e.isCreatureType(EnumCreatureType.MONSTER, false)) {
+        e.setFire(FIRESECONDS);
+        //e.addPotionEffect(new PotionEffect(MobEffects.SPEED, 20 * 9, 1));
+      }
+    }
+    super.onEntityCollidedWithBlock(worldIn, pos, state, entityIn);
+  }
+}
